Post-review changes

+++ stock-service/src/main/java/org/trainee/stockservice/service/StockService.java
@@ -4,6 +4,7 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
+import org.trainee.productservice.enums.EntityType;
 import org.trainee.productservice.exception.EntityNotFoundException;
 import org.trainee.productservice.model.Product;
 import org.trainee.stockservice.dto.StockProductRequest;
@@ -15,8 +16,8 @@ import org.trainee.stockservice.model.StockProduct;
 import org.trainee.stockservice.repository.StockProductRepository;
 import org.trainee.stockservice.repository.StockRepository;
 
+import java.text.MessageFormat;
 import java.util.List;
-import java.util.Optional;
 
 
 @Service
@@ -24,9 +25,8 @@ public class StockService {
     private final StockRepository stockRepository;
     private final StockProductRepository stockProductRepository;
     private RestTemplate restTemplate;
-    String productServiceUrl = "http://localhost:8080/api/v1/product";
-    final String STOCK_ERROR_MESSAGE = "Stock not found!";
-    final String PRODUCT_ERROR_MESSAGE = "Product not found!";
+    String productServiceUrl = "http://localhost:8080/api/v1/product/";
+    final static String STOCK_ERROR_MESSAGE = "Entity with name: {0} and id {1} not found!";
 
     @Autowired
     public StockService(StockRepository stockRepository, StockProductRepository stockProductRepository, RestTemplate restTemplate) {
@@ -39,18 +39,21 @@ public class StockService {
         return stockRepository.findAll().stream().map(StockMapper::stockToResponse).toList();
     }
 
-    public Optional<StockResponse> getStockById(Long id) {
-        return stockRepository.findById(id)
-                .map(StockMapper::stockToResponse);
+    public StockResponse getStockById(Long id) {
+        Stock stock = stockRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("m"));
+        return StockMapper.stockToResponse(stock);
     }
 
     @Transactional
     public StockResponse addProduct(Long stockId, StockProductRequest stockProductRequest) {
+        String stockMessage = MessageFormat.format(STOCK_ERROR_MESSAGE, EntityType.STOCK.name(), stockId);
+        String productMessage = MessageFormat.format(STOCK_ERROR_MESSAGE, EntityType.PRODUCT.name(), stockProductRequest.getProductId());
+
         Stock stock = stockRepository.findById(stockId)
-                .orElseThrow(() -> new EntityNotFoundException(STOCK_ERROR_MESSAGE));
-        Product product = restTemplate.getForObject(productServiceUrl + "/" + stockProductRequest.getProductId(), Product.class);
+                .orElseThrow(() -> new EntityNotFoundException(stockMessage));
+        Product product = restTemplate.getForObject(productServiceUrl + stockProductRequest.getProductId(), Product.class);
         if (product == null) {
-            throw new EntityNotFoundException(PRODUCT_ERROR_MESSAGE);
+            throw new EntityNotFoundException(productMessage);
         }
         StockProduct stockProduct = StockMapper.requestToStockProduct(stockProductRequest);
         stockProduct.setStock(stock);
@@ -62,9 +65,15 @@ public class StockService {
     }
 
     public void removeProduct(Long stockId, Long productId) {
-        stockRepository.findById(stockId)
-                .orElseThrow(() -> new EntityNotFoundException(STOCK_ERROR_MESSAGE));
+        String stockMessage = MessageFormat.format(STOCK_ERROR_MESSAGE, EntityType.STOCK.name(), stockId);
 
+        String productMessage = MessageFormat.format(STOCK_ERROR_MESSAGE, EntityType.PRODUCT.name(), productId);
+        stockRepository.findById(stockId)
+                .orElseThrow(() -> new EntityNotFoundException(stockMessage));
+        Product product = restTemplate.getForObject(productServiceUrl + productId, Product.class);
+        if (product == null) {
+            throw new EntityNotFoundException(productMessage);
+        }
         StockProductKey id = new StockProductKey(stockId, productId);
         stockProductRepository.deleteById(id);
     }
